pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["Deivison Isidoro da Silva <61429017+deivisonisidoro@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2 = "^2.9.6"
pydantic = "^1.10.9"
django = "^4.2.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
email-validator = "^2.0.0.post2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pylint = "^2.17.4"

[tool.poetry.scripts]
lint = "python ./school/infrastructure/school_django/manage.py"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
